Update: move minidata function to its folder
